feat(#29): remove old code and move mocks
