+++ cmd/root.go
@@ -153,15 +153,6 @@ func runKrokCmd(cmd *cobra.Command, args []string) {
 		GlobalTokenKey: krokArgs.server.GlobalTokenKey,
 	}, handlerDeps)
 
-	repoHandler, _ := handlers.NewRepositoryHandler(handlers.Config{
-		Hostname:       krokArgs.server.Hostname,
-		GlobalTokenKey: krokArgs.server.GlobalTokenKey,
-	}, handlers.RepoHandlerDependencies{
-		RepositoryStorer: repoStore,
-		TokenProvider:    tp,
-		Logger:           log,
-	})
-
 	commandHandler, _ := handlers.NewCommandsHandler(handlers.Config{
 		Hostname:       krokArgs.server.Hostname,
 		GlobalTokenKey: krokArgs.server.GlobalTokenKey,
@@ -190,11 +181,10 @@ func runKrokCmd(cmd *cobra.Command, args []string) {
 
 	repoSvcConfig := service.RepositoryServiceConfig{Hostname: krokArgs.server.Hostname}
 	server := server.NewKrokServer(krokArgs.server, server.Dependencies{
-		Logger:            log,
-		Krok:              krokHandler,
-		RepositoryHandler: repoHandler,
-		CommandHandler:    commandHandler,
-		ApiKeyHandler:     apiKeysHandler,
+		Logger:         log,
+		Krok:           krokHandler,
+		CommandHandler: commandHandler,
+		ApiKeyHandler:  apiKeysHandler,
 
 		// TODO: Find a better place?
 		TokenProvider:     tp,
