+++ pkg/server/server.go
@@ -46,11 +46,10 @@ type KrokServer struct {
 
 // Dependencies defines needed dependencies for the krok server.
 type Dependencies struct {
-	Logger            zerolog.Logger
-	Krok              krok.Handler
-	RepositoryHandler providers.RepositoryHandler
-	CommandHandler    providers.CommandHandler
-	ApiKeyHandler     providers.ApiKeysHandler
+	Logger         zerolog.Logger
+	Krok           krok.Handler
+	CommandHandler providers.CommandHandler
+	ApiKeyHandler  providers.ApiKeysHandler
 
 	TokenProvider     providers.TokenProvider
 	RepositoryService repov1.RepositoryServiceServer
